% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdpois.R
\name{app}
\alias{app}
\title{Producing Poisson model data frame}
\usage{
app(x, t, to)
}
\arguments{
\item{x, t}{Numeric vector giving the coordinates of the points to be interpolated.}

\item{to}{an optional set of numeric values specifying where interpolation is to take place}
}
\description{
Producing Poisson model data frame
}
