% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawSurv.R
\name{averageData}
\alias{averageData}
\title{Make average data from a data.frame}
\usage{
averageData(data, newValue = list())
}
\arguments{
\item{data}{A data.frame}

\item{newValue}{Optional list of preset values}
}
\value{
A data.frame
}
\description{
Make average data from a data.frame
}
\examples{
data(pbc,package="survival")
dfSex=averageData(pbc,list(sex=c("m","f")))
dfBili=averageData(pbc,list(bili=c(1,10)))
}
