% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggGam.R
\name{basisFun}
\alias{basisFun}
\title{Show basis function}
\usage{
basisFun(model, which = NULL, ...)
}
\arguments{
\item{model}{An object of class gam}

\item{which}{NULL or one of 1:4}

\item{...}{Further arguments to be passed to plot}
}
\description{
Show basis function
}
\examples{
require(mgcv)
set.seed(1)
x <- seq(0, pi * 2, 0.1)
sin_x <- sin(x)
y <- sin_x + rnorm(n = length(x), mean = 0, sd = sd(sin_x / 2))
data1 <- data.frame(y,x)
model=gam(y~s(x),method="REML")
basisFun(model)
basisFun(model,which=1)
basisFun(model,which=2)
basisFun(model,which=3)
basisFun(model,which=4)
}
