% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggGam.R
\name{call2vars}
\alias{call2vars}
\title{Make list of args with call string}
\usage{
call2vars(string)
}
\arguments{
\item{string}{A character}
}
\value{
A list
}
\description{
Make list of args with call string
}
