% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdpois.R
\name{drawFUData}
\alias{drawFUData}
\title{Draw follow-up plot for laboratory data}
\usage{
drawFUData(
  model,
  data,
  which = c("protime", "platelet", "bili", "albumin", "ast"),
  point = TRUE,
  line = TRUE
)
}
\arguments{
\item{model}{An object of class gam}

\item{data}{A data.frame}

\item{which}{A character vector Names of columns to drae}

\item{point}{logical Whether or not draw point}

\item{line}{logical Whether or not draw line}
}
\description{
Draw follow-up plot for laboratory data
}
\examples{
library(survival)
library(mgcv)
pbcseq$status1 <- as.numeric(pbcseq$status==2) ## deaths
pb <- tdpois(pbcseq) ## conversion
pb$tf <- factor(pb$futime) ## add factor for event time
b <- bam(z ~ tf - 1 + trt + s(sqrt(protime)) + s(platelet) + s(age) + s(bili) +
s(albumin) + s(sqrt(ast)),family=poisson,data=pb,discrete=TRUE,nthreads=2)
data=pbcseq[pbcseq$id==25,]
drawFUData(b,data,which=c("protime","platelet","bili","albumin","ast"))
}
