% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdpois.R
\name{drawFUSurv}
\alias{drawFUSurv}
\title{Draw Kaplan-Meier Survival Curve For Follow-up Data}
\usage{
drawFUSurv(model, data)
}
\arguments{
\item{model}{An object of class gam}

\item{data}{A data.frame}
}
\description{
Draw Kaplan-Meier Survival Curve For Follow-up Data
}
\examples{
library(survival)
library(mgcv)
pbcseq$status1 <- as.numeric(pbcseq$status==2) ## deaths
pb <- tdpois(pbcseq) ## conversion
pb$tf <- factor(pb$futime) ## add factor for event time
b <- bam(z ~ tf - 1 + trt + s(sqrt(protime)) + s(platelet) + s(age) + s(bili) +
s(albumin) + s(sqrt(ast)),family=poisson,data=pb,discrete=TRUE,nthreads=2)
data=pbcseq[pbcseq$id \%in\% c(10,25,66),]
data=pbcseq[pbcseq$id \%in\% c(25),]
drawFUSurv(b,data)
}
