% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawSurv.R
\name{drawSurv}
\alias{drawSurv}
\title{Draw Kaplan-Meier Survival Plot with an object of class 'gam'}
\usage{
drawSurv(model, data, np = 100, timevar = "time", until = NULL, id = list())
}
\arguments{
\item{model}{a fitted gam object as produced by gam()}

\item{data}{A data frame or list containing the values of the model covariates at which predictions are required}

\item{np}{Numeric Number of time intervals}

\item{timevar}{character Name of variable containing time}

\item{until}{Maximum value of time}

\item{id}{Optional labels of preset values}
}
\value{
A ggplot
}
\description{
Draw Kaplan-Meier Survival Plot with an object of class 'gam'
}
\examples{
require(mgcv)
data(pbc,package="survival")
pbc$status1 <- as.numeric(pbc$status==2)
pbc$stage <- factor(pbc$stage)
model = gam(time ~ trt +sex + s(sqrt(protime))+s(platelet)+ s(age)+s(bili)+s(albumin),
   weights=status1, family=cox.ph, data=pbc, method="REML")
dfSex=averageData(pbc,list(sex=c("m","f")))
drawSurv(model,data=dfSex,id=list(sex=c("m","f")))
dfBili=averageData(pbc,list(bili=c(1,10)))
drawSurv(model,data=dfBili,id=list(bili=c(1,10)))
}
