% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggGam.R
\name{formula2vars}
\alias{formula2vars}
\title{Extract independent variable name from formula}
\usage{
formula2vars(formula)
}
\arguments{
\item{formula}{formula}
}
\description{
Extract independent variable name from formula
}
