% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam.Dx.R
\name{gam.Dx}
\alias{gam.Dx}
\title{Draw diagnostic plots for a fitted gam model}
\usage{
gam.Dx(
  b,
  type = c("deviance", "pearson", "response"),
  rep = 0,
  level = 0.9,
  select = NULL
)
}
\arguments{
\item{b}{a fitted gam object as produced by gam().}

\item{type}{type of residuals}

\item{rep}{arguments passed to qq.gam()}

\item{level}{arguments passed to qq.gam()}

\item{select}{numeric choices of plots}
}
\description{
Draw diagnostic plots for a fitted gam model
}
\examples{
require(mgcv)
data(pbc,package="survival")
pbc$status1 <- as.numeric(pbc$status==2)
pbc$stage <- factor(pbc$stage)
b = gam(time ~ trt +sex + s(sqrt(protime))+s(platelet)+ s(age)+s(bili)+s(albumin),
   weights=status1, family=cox.ph, data=pbc, method="REML")
gam.Dx(b)
gam.Dx(b,select=4)
}
