% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamCheck.R
\name{gamCheck}
\alias{gamCheck}
\title{Perform diagnostics for a fitted gam model}
\usage{
gamCheck(b, k.sample = 5000, k.rep = 200, print = TRUE)
}
\arguments{
\item{b}{A fitted gam object as produced by gam().}

\item{k.sample}{Newmeric Above this k testing uses a random sub-sample of data.}

\item{k.rep}{Numeric How many re-shuffles to do to get p-value for k testing.}

\item{print}{logical Whether or not print result}
}
\description{
Perform diagnostics for a fitted gam model
}
\examples{
library(mgcv)
set.seed(0)
dat <- gamSim(1,n=200)
b<-gam(y~s(x0)+s(x1)+s(x2)+s(x3),data=dat)
gamCheck(b)
}
