% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggGam.R
\name{ggGam}
\alias{ggGam}
\title{Draw a ggplot with an object of class gam}
\usage{
ggGam(
  model,
  select = NULL,
  point = TRUE,
  se = TRUE,
  by = NULL,
  scales = "free_x",
  type = NULL
)
}
\arguments{
\item{model}{An object of class gam}

\item{select}{numeric Choices of dependent variables to plot}

\item{point}{logical Whether or not draw point}

\item{se}{logical Whether or not draw confidence interval}

\item{by}{NULL or character optional name of factor variable}

\item{scales}{Should scales be fixed ("fixed"), free ("free"), or free in one dimension ("free_x", "free_y")?}

\item{type}{character type argument to be passed to predict.gam}
}
\description{
Draw a ggplot with an object of class gam
}
\examples{
require(mgcv)
model <- gam(mpg ~ s(wt), data = mtcars, method = "REML")
plot(model,shift=coef(model)[1],pages=1,all.terms=TRUE,shade=TRUE,seWithMean=TRUE,residuals=TRUE)
ggGam(model)
mtcars$am=factor(mtcars$am,labels=c("automatic","manual"))
model <- gam(mpg ~ s(wt)+am, data = mtcars, method = "REML")
plot(model,shift=coef(model)[1],pages=1,all.terms=TRUE,shade=TRUE,seWithMean=TRUE,residuals=TRUE)
ggGam(model)
ggGam(model,by=am)
model <- gam(mpg ~ s(wt,by=am)+am, data = mtcars, method = "REML")
plot(model,shift=coef(model)[1],pages=1,all.terms=TRUE,shade=TRUE,seWithMean=TRUE,residuals=TRUE)
ggGam(model)
ggGam(model,by=am)
ggGam(model,by=am,point=FALSE)
ggGamCat(model)
data(mpg,package="gamair")
model <- gam(hw.mpg ~ s(weight, fuel, bs = "fs"),data = mpg,method = "REML")
ggGam(model)
model2 <- gam(hw.mpg ~ s(weight) + s(length) + s(price) + fuel + drive + style,
   data=mpg, method="REML")
plot(model2,shift=coef(model)[1],pages=1,all.terms=TRUE,shade=TRUE,seWithMean=TRUE,residuals=TRUE)
ggGam(model2,se=TRUE)
ggGam(model2,se=TRUE,by=fuel,select=1)
ggGam(model2,se=FALSE,by=style)
ggGam(model2,se=FALSE,by=drive,point=FALSE)
}
