% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggGam.R
\name{ggGamCat}
\alias{ggGamCat}
\title{Draw a ggplot summarizing categorical variables with an object of class gam}
\usage{
ggGamCat(model, select = NULL, scales = "free_x", type = NULL)
}
\arguments{
\item{model}{An object of class gam}

\item{select}{numeric Choices of dependent categorical variables to plot}

\item{scales}{Should scales be fixed ("fixed"), free ("free"), or free in one dimension ("free_x", "free_y")?}

\item{type}{character type argument to be passed to predict.gam}
}
\description{
Draw a ggplot summarizing categorical variables with an object of class gam
}
\examples{
require(mgcv)
data(Wage,package="ISLR")
model=gam(wage~s(age,by=education)+education+jobclass,data=Wage,method="REML")
plot(model,shift=coef(model)[1],pages=1,all.terms=TRUE,shade=TRUE,
    seWithMean=TRUE,residuals=TRUE)
ggGamCat(model)
}
