% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam.Dx.R
\name{qqgam}
\alias{qqgam}
\title{QQ plots for gam model residuals}
\usage{
qqgam(
  object,
  rep = 0,
  level = 0.9,
  s.rep = 10,
  type = c("deviance", "pearson", "response")
)
}
\arguments{
\item{object}{a fitted gam object as produced by gam()}

\item{rep}{How many replicate datasets to generate to simulate quantiles of the residual distribution. 0 results in an efficient simulation free method for direct calculation, if this is possible for the object family.}

\item{level}{If simulation is used for the quantiles, then reference intervals can be provided for the QQ-plot, this specifies the level. 0 or less for no intervals, 1 or more to simply plot the QQ plot for each replicate generated.}

\item{s.rep}{how many times to randomize uniform quantiles to data under direct computation.}

\item{type}{what sort of residuals should be plotted?}
}
\value{
A ggplot
}
\description{
QQ plots for gam model residuals
}
