% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdpois.R
\name{stepfun2df}
\alias{stepfun2df}
\title{Extract x and coordinate with stepfun}
\usage{
stepfun2df(x)
}
\arguments{
\item{x}{An object of class stepfun}
}
\value{
A data.frame
}
\description{
Extract x and coordinate with stepfun
}
