% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdpois.R
\name{tdpois}
\alias{tdpois}
\title{Convert data.frame into a data.frame of artificial Poisson data with the interpolated Time varying covariate values}
\usage{
tdpois(
  dat,
  event = "z",
  et = "futime",
  t = "day",
  status = "status1",
  id = "id"
)
}
\arguments{
\item{dat}{A data.frame}

\item{event}{Name for Poisson response}

\item{et}{Name of event time}

\item{t}{Name of observation time}

\item{status}{Numeric 1 for death 0 otherwise}

\item{id}{Patient id}
}
\value{
A data.frame of artificial Poisson data
}
\description{
Convert data.frame into a data.frame of artificial Poisson data with the interpolated Time varying covariate values
}
\examples{
library(survival)
pbcseq$status1 <- as.numeric(pbcseq$status==2) ## deaths
pb <- tdpois(pbcseq) ## conversion
}
