% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggGam.R
\name{ggGam}
\alias{ggGam}
\title{Draw a ggplot with an object of class gam}
\usage{
ggGam(
  model,
  select = NULL,
  point = FALSE,
  se = TRUE,
  by = NULL,
  byauto = TRUE,
  scales = "free_x"
)
}
\arguments{
\item{model}{An object of class gam}

\item{select}{numeric Choices of dependent variables to plot}

\item{point}{logical Whether or not draw point}

\item{se}{logical Whether or not draw confidence interval}

\item{by}{character optional name of factor variable}

\item{byauto}{logical Whether or not select categorical variables automatically}

\item{scales}{Should scales be fixed ("fixed"), free ("free"), or free in one dimension ("free_x", "free_y")?}
}
\description{
Draw a ggplot with an object of class gam
}
\examples{
require(mgcv)
mtcars$am=factor(mtcars$am,labels=c("automatic","manual"))
model <- gam(mpg ~ s(wt,by=am)+am, data = mtcars, method = "REML")
plot(model,shift=coef(model)[1],pages=1,all.terms=TRUE,shade=TRUE,seWithMean=TRUE,residuals=TRUE)
ggGam(model)
ggGam(model,point=TRUE)
}
