% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggGam.R
\name{ggGamCat}
\alias{ggGamCat}
\title{Draw a ggplot summarizing categorical variables with an object of class gam}
\usage{
ggGamCat(model, scales = "free_x")
}
\arguments{
\item{model}{An object of class gam}

\item{scales}{Should scales be fixed ("fixed"), free ("free"), or free in one dimension ("free_x", "free_y")?}
}
\description{
Draw a ggplot summarizing categorical variables with an object of class gam
}
\examples{
require(mgcv)
data(Wage,package="ISLR")
model=gam(wage~s(age,by=education)+education+jobclass,data=Wage,method="REML")
plot(model,shift=coef(model)[1],pages=1,all.terms=TRUE,shade=TRUE,
    seWithMean=TRUE,residuals=TRUE)
ggGamCat(model)
}
